+++ src/disk_object.cpp
@@ -25,15 +25,19 @@ void DiskObject::reset() {
 void DiskObject::render(DemoApplication *app) {
     DemoObject::render(app);
 
-    app->drawDisk(m_body.p_x, m_body.p_y, m_body.theta, m_radius);
+    app->drawDisk(
+        (float)m_body.p_x,
+        (float)m_body.p_y,
+        (float)m_body.theta,
+        m_radius);
 }
 
 void DiskObject::process(float dt, DemoApplication *app) {
     DemoObject::process(dt, app);
 }
 
-void DiskObject::configure(float radius, float density) {
-    m_radius = radius;
+void DiskObject::configure(double radius, double density) {
+    m_radius = (float)radius;
 
     m_body.m = density * (double)ysMath::Constants::PI * radius * radius;
     m_body.I = 0.5 * m_body.m * radius * radius;
@@ -48,7 +52,7 @@ void DiskObject::onClick(double x, double y, ClickEvent *clickEvent) {
     m_body.worldToLocal(x, y, &lx, &ly);
 
     clickEvent->clicked = false;
-    if (lx * lx + ly * ly > m_radius * m_radius) return;
+    if (lx * lx + ly * ly > (double)m_radius * m_radius) return;
     
     clickEvent->body = &m_body;
     clickEvent->clicked = true;
